+++ index.d.ts
@@ -212,7 +212,10 @@ export class CommanderError extends Error {
     description: string;
     required: boolean;
     variadic: boolean;
-  
+    defaultValue?: any;
+    defaultValueDescription?: string;
+    argChoices?: string[];
+
     /**
      * Initialize a new command argument with the given name and description.
      * The default is that the argument is required, and you can explicitly
@@ -259,12 +262,13 @@ export class CommanderError extends Error {
     optional: boolean; // A value is optional when the option is specified.
     variadic: boolean;
     mandatory: boolean; // The option must have a value after parsing, which usually means it must be specified on command line.
-    optionFlags: string;
     short?: string;
     long?: string;
     negate: boolean;
     defaultValue?: any;
     defaultValueDescription?: string;
+    presetArg?: unknown;
+    envVar?: string;
     parseArg?: <T>(value: string, previous: T) => T;
     hidden: boolean;
     argChoices?: string[];
@@ -446,14 +450,12 @@ export class CommanderError extends Error {
   // The source is a string so author can define their own too.
   export type OptionValueSource = LiteralUnion<'default' | 'config' | 'env' | 'cli' | 'implied', string> | undefined;
 
-  export interface OptionValues {
-    [key: string]: unknown;
-  }
+  export type OptionValues = Record<string, unknown>;
   
   export class Command<Args extends any[] = [], Opts extends OptionValues = {}> {
     args: string[];
     processedArgs: Args;
-    commands: CommandUnknownOpts[];
+    readonly commands: readonly CommandUnknownOpts[];
     readonly options: readonly Option[];
     readonly registeredArguments: readonly Argument[];
     parent: CommandUnknownOpts | null;
@@ -469,7 +471,10 @@ export class CommanderError extends Error {
      * You can optionally supply the  flags and description to override the defaults.
      */
     version(str: string, flags?: string, description?: string): this;
-  
+    /**
+     * Get the program version.
+     */
+    version(): string | undefined;
     /**
      * Define a command, implemented using an action handler.
      *
@@ -681,45 +686,21 @@ export class CommanderError extends Error {
     action(fn: (...args: [...Args, Opts, this]) => void | Promise<void>): this;
 
     /**
-     * Define option with `flags`, `description` and optional
-     * coercion `fn`.
+     * Define option with `flags`, `description`, and optional argument parsing function or `defaultValue` or both.
      *
-     * The `flags` string contains the short and/or long flags,
-     * separated by comma, a pipe or space. The following are all valid
-     * all will output this way when `--help` is used.
+     * The `flags` string contains the short and/or long flags, separated by comma, a pipe or space. A required
+     * option-argument is indicated by `<>` and an optional option-argument by `[]`.
      *
-     *     "-p, --pepper"
-     *     "-p|--pepper"
-     *     "-p --pepper"
+     * See the README for more details, and see also addOption() and requiredOption().
      *
      * @example
-     * ```
-     * // simple boolean defaulting to false
-     *  program.option('-p, --pepper', 'add pepper');
      *
-     *  --pepper
-     *  program.pepper
-     *  // => Boolean
-     *
-     *  // simple boolean defaulting to true
-     *  program.option('-C, --no-cheese', 'remove cheese');
-     *
-     *  program.cheese
-     *  // => true
-     *
-     *  --no-cheese
-     *  program.cheese
-     *  // => false
-     *
-     *  // required argument
-     *  program.option('-C, --chdir <path>', 'change the working directory');
-     *
-     *  --chdir /tmp
-     *  program.chdir
-     *  // => "/tmp"
-     *
-     *  // optional argument
-     *  program.option('-c, --cheese [type]', 'add cheese [marble]');
+     * ```js
+     * program
+     *     .option('-p, --pepper', 'add pepper')
+     *     .option('-p, --pizza-type <TYPE>', 'type of pizza') // required option-argument
+     *     .option('-c, --cheese [CHEESE]', 'add extra cheese', 'mozzarella') // optional option-argument with default
+     *     .option('-t, --tip <VALUE>', 'add tip to purchase cost', parseFloat) // custom parse function
      * ```
      *
      * @returns `this` command for chaining
@@ -729,9 +710,9 @@ export class CommanderError extends Error {
     option<S extends string, DefaultT extends string | boolean | string[] | []>(
         usage: S, description?: string, defaultValue?: DefaultT): Command<Args, InferOptions<Opts, S, DefaultT, undefined, false>>;
     option<S extends string, T>(
-        usage: S, description: string, fn: (value: string, previous: T) => T): Command<Args, InferOptions<Opts, S, undefined, T, false>>;
+        usage: S, description: string, parseArg: (value: string, previous: T) => T): Command<Args, InferOptions<Opts, S, undefined, T, false>>;
     option<S extends string, T>(
-        usage: S, description: string, fn: (value: string, previous: T) => T, defaultValue?: T): Command<Args, InferOptions<Opts, S, T, T, false>>;
+        usage: S, description: string, parseArg: (value: string, previous: T) => T, defaultValue?: T): Command<Args, InferOptions<Opts, S, T, T, false>>;
       
     /**
      * Define a required option, which must have a value after parsing. This usually means
@@ -744,9 +725,9 @@ export class CommanderError extends Error {
     requiredOption<S extends string, DefaultT extends string | boolean | string[]>(
         usage: S, description?: string, defaultValue?: DefaultT): Command<Args, InferOptions<Opts, S, DefaultT, undefined, true>>;
     requiredOption<S extends string, T>(
-        usage: S, description: string, fn: (value: string, previous: T) => T): Command<Args, InferOptions<Opts, S, undefined, T, true>>;
+        usage: S, description: string, parseArg: (value: string, previous: T) => T): Command<Args, InferOptions<Opts, S, undefined, T, true>>;
     requiredOption<S extends string, T, D extends T>(
-        usage: S, description: string, fn: (value: string, previous: T) => T, defaultValue?: D): Command<Args, InferOptions<Opts, S, D, T, true>>;
+        usage: S, description: string, parseArg: (value: string, previous: T) => T, defaultValue?: D): Command<Args, InferOptions<Opts, S, D, T, true>>;
       
     /**
      * Factory routine to create a new unattached option.
@@ -920,7 +901,7 @@ export class CommanderError extends Error {
   
     description(str: string): this;
     /** @deprecated since v8, instead use .argument to add command argument with description */
-    description(str: string, argsDescription: {[argName: string]: string}): this;
+    description(str: string, argsDescription: Record<string, string>): this;
     /**
      * Get the description.
      */
@@ -1017,7 +998,7 @@ export class CommanderError extends Error {
     /**
      * Get the executable search directory.
      */
-    executableDir(): string;
+    executableDir(): string | null;
   
     /**
      * Output help information for this command.